requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asteval==1.0.5
astroid==3.3.5
astunparse==1.6.3
black==24.10.0
blinker==1.8.2
brightway2==2.4.6
Brotli==1.1.0
build==1.2.2.post1
bw-migrations==0.2
bw2analyzer==0.10
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
CacheControl==0.14.1
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crashtest==0.4.1
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
docopt==0.6.2
dulwich==0.21.7
ecoinvent_interface==2.5
eight==1.0.1
et-xmlfile==1.1.0
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
flexcache==0.3
flexparser==0.3.1
fonttools==4.54.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflate64==1.0.0
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
more-itertools==10.5.0
mrio-common-metadata==0.2.1
msgpack==1.1.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
Pint==0.24.3
pkginfo==1.11.2
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psutil==6.0.0
ptyprocess==0.7.0
py7zr==0.22.0
pybcj==1.0.2
pycodestyle==2.12.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyecospold==4.0.0
pyflakes==3.2.0
pylint==3.3.1
pyparsing==3.1.4
pyppmd==1.1.0
PyPrind==2.11.3
pyproject_hooks==1.2.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytest==8.3.3
pytest-qt==4.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
pyxlsb==1.0.10
pyzstd==0.16.1
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.13.1
shellingham==1.5.4
shiboken6==6.8.0.2
six==1.16.0
smmap==5.0.1
stats_arrays==0.7
tabulate==0.9.0
tenacity==9.0.0
texttable==1.7.0
tomlkit==0.13.2
tqdm==4.66.5
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
unicodecsv==0.14.1
Unidecode==1.3.8
urllib3==2.2.3
virtualenv==20.27.1
voluptuous==0.15.2
Werkzeug==3.0.4
wheel==0.45.0
Whoosh==2.7.4
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
zipp==3.20.2
